html/browsers/history/the-location-interface/location-protocol-setter.html

<!doctype html>
<title>Set location.protocol to broken schemes</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
  let broken = [
    'x\x20',
    'x\x00',
    '\x00',
    ':',
    '†',
    'a†',
    'a\x0A',
    'a+-.\x0A',
    '\x7F',
    '\xFF'
  ]
  ;broken.forEach((val) => {
    test(() => {
      assert_throws(new TypeError, () => { location.protocol = val })
    }, encodeURI(val) + " (percent-encoded) is not a scheme")
  })
</script>
